Add initial build support for DSO Nano
There are lots of empty placeholders in the code, so not
much can work at this point.

However, it builds, and what is missing is identified in
DS203/nanoglue.c and just needs to be filled in...

This implementation reuses code in the DSO Nano open-source
firmware LIB, and assumes this to be installed at 0x8007800
(dual-boot configuration).

The linker script places the code at 0x08014C00, just after the BenF
3.64 binary. This makes it overwrite the "slot" of the open-source
dual-boot Benf+fft - there is simply not enough space for keeping
both of them.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>